Initial commit.
Did a lot for this first commit. The I grab the leapmotion data in a
tight polling loop.

Key presses and mouse emulation is done through the xcb-xtest library.

The program is capable of scrolling when finger is detected going
clockwise or counter clockwise. It scrolls more or less depending on the
radius of the circle.

The program is also capable of swiping between workspaces with a
vertical swipe.